pyproject.toml
[tool.poetry]
name = "mywill_scanner"
version = "0.1.0"
description = ""
authors = ["Ameize4 <ameize4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
certifi = "^2020.12.5"
chardet = "^4.0.0"
PyYAML = "^5.4.1"
web3 = "^5.17.0"
urllib3 = "^1.26.3"
python-bitcoinrpc = "^1.0"
PyPubSub = "^4.0.3"
psycopg2-binary = "^2.8.6"
pika = "^1.2.0"
tronapi = {git = "https://github.com/MyWishPlatform/tron-api-python.git"}
sqlalchemy = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

